% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binar_search.R
\name{binar_search}
\alias{binar_search}
\title{Binary search}
\usage{
binar_search(X, x3, l = 1, h = length(X))
}
\arguments{
\item{X}{A numeric vector, sorted in increasing order.}

\item{x3}{A value of interest.}

\item{l}{The highest possible lower index that surrounds a value of interest from left (bottom). Default value 1.}

\item{h}{The lowest possible higher index that surrounds a value of interest from right (top).Default value is the length of X.}
}
\value{
A filtered data frame.
}
\description{
This function helps to find two adjacent indeces of a numeric vector, sorted in increasing order, that surround a value of interest.
}
\note{
This function drops points that lie strictly outside of a given polygon. If a point lies on a relative interior of an edge of the polygon (provided that polygon is open) or point is a vertex of the polygon, then it will be retained.
}
\examples{
binar_search(X=1:100,x3=55.5)
}
\author{
Liudas Daumantas
}
