% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block3D.R
\name{blok3d}
\alias{blok3d}
\title{Draw hespdiv polygons in 3D space}
\usage{
blok3d(
  obj,
  height = "mean",
  color.seed = 1,
  lines = TRUE,
  pnts.col = NULL,
  obs = TRUE
)
}
\arguments{
\item{obj}{An object of the hespdiv class}

\item{height}{A character vector with a default value of 'mean'. Which
information from the "poly.stats" data frame do you want to encode as the
height of the polygons? Options:
"mean", "sd", "best", "z.score", "str.best", "str.z.score","rank". Multiple
values are allowed.}

\item{color.seed}{An integer that controls the colors of the polygons. Change
it to a different number if you want to get a different set of colors.}

\item{lines}{A Boolean value. Do you want split-lines to be displayed over
the top of the polygons?}

\item{pnts.col}{A character or numeric vector. Color codes to be used for
displaying observations.}

\item{obs}{A Boolean value. Do you want observations to be displayed over the top of
the polygons?}
}
\value{
No return value. Called for the side effect of producing a 3D plot
using the \pkg{rgl} graphics engine.
}
\description{
This function visualizes HespDiv polygons in 3D space. The height axis
corresponds to a chosen column from the "poly.stats" data frame.
}
\details{
The function opens an rgl device for each column selected from the 'poly.stats'
data frame.

Visualizing the column values of "poly.stats" as polygon height can provide
insight into the spatial heterogeneity of the analyzed data and its spatial
hierarchical structure.

Additionally, the 'height = rank' option provides a more intuitive way to
understand the location of each polygon compared to the poly_scheme function.

As polygons of higher rank are displayed on top of the lower rank polygons,
a higher rank polygon might obscure the view. For this reason, the
\code{polypop(obj,height)} function with the same arguments can be used to
interactively select unwanted polygons and remove them from a plot.
}
\seealso{
Other HespDiv visualization options: 
\code{\link{create_gif}()},
\code{\link{dendro}()},
\code{\link{plot.nullhespdiv}()},
\code{\link{plot_cs_hsa}()},
\code{\link{plot_hespdiv}()},
\code{\link{plot_hsa}()},
\code{\link{plot_hsa_q}()},
\code{\link{poly_scheme}()},
\code{\link{polypop}()}
}
\author{
Liudas Daumantas
}
\concept{HespDiv visualization options}
