% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_base.R
\name{change_base}
\alias{change_base}
\title{Change the basal subdivision in hsa}
\usage{
change_base(obj, id)
}
\arguments{
\item{obj}{A \code{hsa} class object.}

\item{id}{An index of an alternative subdivision to be used as a new basal
subdivision.}
}
\value{
The \code{hsa} class object with a new basal subdivision.
}
\description{
This function allows you to select a new basal subdivision
from the results of hespdiv sensitivity analysis by specifying its ID. It
provides a convenient way to switch between different basal subdivisions.
You can identify a more stable subdivision alternative by plotting the
hespdiv sensitivity analysis results using the \code{plot_hsa} function.
By selecting a new basal subdivision, you can observe how it affects the
results of polygon object cross-comparison and the stability of hespdiv
clusters and polygons.
}
\seealso{
Other {functions for hespdiv sensitivity analysis}: 
\code{\link{hsa_detailed}()},
\code{\link{hsa_quant}()},
\code{\link{hsa}()},
\code{\link{plot_hsa_q}()},
\code{\link{plot_hsa}()}
}
\author{
Liudas Daumantas
}
\concept{{functions for hespdiv sensitivity analysis}}
