% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gif.R
\name{create_gif}
\alias{create_gif}
\title{Create a gif of blok3D}
\usage{
create_gif(output_file, frames = 90, angle_per_frame = 5, fps = 10)
}
\arguments{
\item{output_file}{name and path of output file}

\item{frames}{integer. Number of frames.}

\item{angle_per_frame}{numeric. Angle to rotate per frame in degrees.}

\item{fps}{integer. Frames per second.}
}
\value{
No return value. Called for the side effect of saving
a GIF of a rotating 3D plot.
}
\description{
Functions creates a gif of rotating polygons displayed in a currently actively rgl device produced with blok3D
}
\note{
You can adjust the size of rgl device window to control the size of gif.
}
\seealso{
Other HespDiv visualization options: 
\code{\link{blok3d}()},
\code{\link{dendro}()},
\code{\link{plot.nullhespdiv}()},
\code{\link{plot_cs_hsa}()},
\code{\link{plot_hespdiv}()},
\code{\link{plot_hsa}()},
\code{\link{plot_hsa_q}()},
\code{\link{poly_scheme}()},
\code{\link{polypop}()}
}
\author{
Liudas Daumantas
}
\concept{HespDiv visualization options}
