% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_comp.R
\name{cross_comp}
\alias{cross_comp}
\title{Calculate a polygon-object cross-comparison matrix}
\usage{
cross_comp(obj)
}
\arguments{
\item{obj}{A \code{hespdiv} object.}
}
\value{
A numeric matrix containing pairwise comparison values among the
\code{hespdiv} polygon objects stored in \code{obj$poly.obj}.
}
\description{
Computes a cross-comparison matrix for polygon objects from a \code{hespdiv}
object. The matrix quantifies similarity or dissimilarity among polygon
objects and can be used for further analyses, such as clustering, either
directly or after transformation.
}
\details{
The \code{cross_comp()} function uses the \code{compare.f} function from
\code{obj$call.info$Call_ARGS} to perform pairwise comparisons of
\code{hespdiv} polygon objects stored in \code{obj$poly.obj}. The result is
a cross-comparison matrix.
}
\note{
Polygon cross-comparison is currently not available for the \code{"pielou"}
method. It is also not supported for custom methods whose \code{compare.f}
function relies on variables from environments other than the function's own
arguments.
}
\seealso{
Other functions for hespdiv results post-processing: 
\code{\link{hsa}()},
\code{\link{hsa_detailed}()},
\code{\link{hsa_quant}()},
\code{\link{hsa_sample_constrained}()},
\code{\link{nulltest}()},
\code{\link{taxon_effect}()}
}
\author{
Liudas Daumantas
}
\concept{functions for hespdiv results post-processing}
