% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depend_splits.R
\name{depend_splits}
\alias{depend_splits}
\title{Identify dependent split-lines and polygons}
\usage{
depend_splits(obj, id)
}
\arguments{
\item{obj}{hespdiv object}

\item{id}{id of a split-line}
}
\value{
a list:
\enumerate{
\item Dependent split-line IDs, if any
\item Dependent polygon IDs
}
}
\description{
For a provided split-line id, function returns a list of
all offspring polygons and split-lines, if there are any.
}
