% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binar_search.R
\name{.binar_search}
\alias{.binar_search}
\title{Binary search}
\usage{
.binar_search(X, x3, l = 1, h = length(X))
}
\arguments{
\item{X}{A numeric vector, sorted in increasing order.}

\item{x3}{A value of interest.}

\item{l}{The highest possible lower index that surrounds a value of interest from left (bottom). Default value 1.}

\item{h}{The lowest possible higher index that surrounds a value of interest from right (top).Default value is the length of X.}
}
\value{
Two adjacent X vector indeces of values that surrounds a value of interest.
}
\description{
This function helps to find two adjacent indeces of a numeric vector, sorted in increasing order, that surround a value of interest.
}
\note{
If the value of interest (\code{x3}) is equal to some value of X, than the index of this value will be returned as a higher index.
Function is not intended to be used in cases of duplicated values.
}
\examples{
# simple case
.binar_search(X=1:100,x3=55.5)
#when x3 is equal to some value of X, index with x3 value in the returned vector will be the higher one.
.binar_search(X=1:100,x3=50)
}
\author{
Liudas Daumantas
}
