% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_on_line.R
\name{.pt_on_line}
\alias{.pt_on_line}
\title{Find a point on a line}
\usage{
.pt_on_line(x1, x2, y1, y2, x3 = NULL, y3 = NULL)
}
\arguments{
\item{x1}{An x coordinate of the start of a line.}

\item{x2}{An x coordinate of the end of a line.}

\item{y1}{An y coordinate of the start of a line.}

\item{y2}{An y coordinate of the end of a line.}

\item{x3}{An x coordinates, where missing y coordinates on a line must be found.}

\item{y3}{An y coordinates, where missing x coordinates on a line must be found.}
}
\value{
A missing x or y coordinates on a line that joins (x1,y1) and (x2,y2) points.
}
\description{
This function helps to find a missing coordinate (x or y) of a point on a straight line that joins two provided points.
}
\examples{
#A line between (3,0) & (1,-1) points
plot(c(3,1),c(0,-1),type='l')
#Finding points at x = 1, 1.5, 2, 2.5, 3
x <- seq(1,3,0.5)
y <- pt_on_line(x1 = 3,x2 = 1,y1 = 0,y2 = -1,x3 = x)
points(x,y,col=2,pch=19)
}
\author{
Liudas Daumantas
}
