% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data points that lie inside a polygon}
\usage{
get_data(polygon, data)
}
\arguments{
\item{polygon}{A data frame of 2 columns (x,y), containing coordinates of polygon vertices}

\item{data}{A data frame, containing "X" and "Y" columns. Other columns may also be present.}
}
\value{
A filtered data frame
}
\description{
Function allows to extract points from a provided dataset that lie inside a given polygon
}
\note{
This function drops points that lie strictly outside of a given polygon. If a point lies on a relative interior of an edge
of polygon (if polygon is open) or point is a vertex of polygon, than it will be retained.
}
\references{
Uses the C function InPoly(). InPoly is Copyright (c) 1998 by Joseph O'Rourke. It may be freely redistributed in its entirety provided that this copyright notice is not removed.
}
\author{
Liudas Daumantas
}
