% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nullhespdiv.R
\name{plot.nullhespdiv}
\alias{plot.nullhespdiv}
\title{Plot a nullhespdiv object}
\usage{
\method{plot}{nullhespdiv}(x, ...)
}
\arguments{
\item{x}{A \code{nullhespdiv} object.}

\item{...}{Additional arguments passed to \code{graphics::boxplot()}.}
}
\value{
No return value. Called for the side effect of plotting a
\code{nullhespdiv} object.
}
\description{
Plot method for a \code{nullhespdiv} object.
}
\examples{
test_results <- nulltest(example_hespdiv)
plot(test_results)
}
\seealso{
Other HespDiv visualization options: 
\code{\link{blok3d}()},
\code{\link{create_gif}()},
\code{\link{dendro}()},
\code{\link{plot_cs_hsa}()},
\code{\link{plot_hespdiv}()},
\code{\link{plot_hsa}()},
\code{\link{plot_hsa_q}()},
\code{\link{poly_scheme}()},
\code{\link{polypop}()}
}
\author{
Liudas Daumantas
}
\concept{HespDiv visualization options}
