% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hespdiv.R
\name{plot_hespdiv}
\alias{plot_hespdiv}
\title{Plot hespdiv results}
\usage{
plot_hespdiv(
  obj,
  type = "color",
  n.loc = FALSE,
  performance = TRUE,
  legend_title = NULL,
  title = NULL,
  subtitle = NULL,
  pnts.col = NULL,
  seed = 10
)
}
\arguments{
\item{obj}{A hespdiv object.}

\item{type}{A character. Either "width" or "color" (default "color").
Determines whether quality of split-lines is expressed by line width or
color.}

\item{n.loc}{A Boolean value. Would you like to visualize the number of observations
at each location? Only possible, when there are localities with more than
one observation. If the type is 'color,' the number of observations is
expressed through point sizes. Otherwise, they are expressed using color in
a logarithmic scale.}

\item{performance}{logical. TRUE - display split-line performance,
FALSE - rank. Displaying rank makes the spatial dendrogram clearer.}

\item{legend_title}{A character value that indicates the title of the legend
for the split-lines. The default is built according to the method information
available in "obj$call.info".}

\item{title}{A character that indicates the title of the plot.}

\item{subtitle}{A character that indicates the subtitle of the plot.}

\item{pnts.col}{A character or numeric vector providing color codes
for data points.}

\item{seed}{An integer value that indicates seed used to randomize the colors
of the split-lines. Only meaningful, when argument \code{type = "width"}.
Try setting a different value, if colors of parallel split-lines or nearby
labels look too similar or to increase the general appeal of the graph.}
}
\value{
A ggplot object.
}
\description{
This function is used to plot the results obtained with the \code{hespidv}
function. The plot showcases subdivisions of the study area by split-lines,
visualizing their performances or rank with colors or line widths. Additionally,
it can display the spatial distribution of observations and number of
observations in each location.
}
\details{
The return ggplot object can be edited as any other ggplot objects
by removing undesired elements, changing theme or overlying the plot
with additional elements.
}
\examples{
plot_hespdiv(example_hespdiv)
plot_hespdiv(example_hespdiv, type = "width")
plot_hespdiv(example_hespdiv, n.loc = TRUE)
}
\seealso{
Other HespDiv visualization options: 
\code{\link{blok3d}()},
\code{\link{create_gif}()},
\code{\link{dendro}()},
\code{\link{plot.nullhespdiv}()},
\code{\link{plot_cs_hsa}()},
\code{\link{plot_hsa}()},
\code{\link{plot_hsa_q}()},
\code{\link{poly_scheme}()},
\code{\link{polypop}()}
}
\author{
Liudas Daumantas
}
\concept{HespDiv visualization options}
