% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hsa_q.R
\name{plot_hsa_q}
\alias{plot_hsa_q}
\title{Plot the stability of hespdiv clusters}
\usage{
plot_hsa_q(obj, hist = FALSE)
}
\arguments{
\item{obj}{The output of \code{hsa_quant} (an object of class \code{hsa_quant}).}

\item{hist}{A Boolean value. If FALSE, EPDFs obtained with \code{plot(density(jac.sim))}
will be displayed instead of histograms.}
}
\value{
None
}
\description{
This function visualizes the stability of basal subdivision
clusters obtained from \code{hsa_quant}
}
\details{
The stability of each basal cluster is revealed by the distribution
of Jaccard similarity values with the 'analogues' clusters found in alternative
hespdiv subdivisions. For example, a unimodal distribution with a peak at
high similarity values (>0.8) indicates that the basal hespdiv cluster is
stable, even if the polygon boundaries are not. This situation may arise when
there is indeed a spatial structure within the data, but there are also wide
gaps between sampled regions (or more generally when there is limited spatial
data coverage). A unimodal distribution with a peak at medium values
(0.4-0.6) and a tail to higher values could also indicate a more persistent
spatial structure. On the other hand, a single peak at low values (<0.4)
indicates low cluster stability (e.g., bioregion does not exist). Finally,
uniform, bimodal, or other more complex distributions may indicate that the
stability and existence of the corresponding basal cluster depend on the
parameters used in alternative hespdiv calls.
}
\seealso{
Other {functions for hespdiv sensitivity analysis}: 
\code{\link{change_base}()},
\code{\link{hsa_detailed}()},
\code{\link{hsa_quant}()},
\code{\link{hsa}()},
\code{\link{plot_hsa}()}

Other {HespDiv visualization options}: 
\code{\link{blok3d}()},
\code{\link{plot_hespdiv}()},
\code{\link{plot_hsa}()},
\code{\link{poly_scheme}()}

Other {functions to evaluate hesdpiv cluster stability}: 
\code{\link{hsa_quant}()}
}
\author{
Liudas Daumantas
}
\concept{{HespDiv visualization options}}
\concept{{functions for hespdiv sensitivity analysis}}
\concept{{functions to evaluate hesdpiv cluster stability}}
