% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_scheme.R
\name{poly_scheme}
\alias{poly_scheme}
\title{Schematic plot of hespdiv polygons}
\usage{
poly_scheme(obj, segment = TRUE, id = TRUE, seed = 1)
}
\arguments{
\item{obj}{A hespdiv object.}

\item{segment}{A Boolean value. Display the punctuated lines joining the
polygon centroids with the split-lines?}

\item{id}{A Boolean value. Display the IDs of polygons?}

\item{seed}{An integer value that determines the random set of colors used
in visualization of split-lines and polygons.}
}
\value{
A ggplot object.
}
\description{
This function generates a schematic visualization of subdivided territory.
It highlights the location of each polygon by displaying their centroids,
ID labels, and punctuated lines that connect the polygon centroids with the
split-lines that created them. This visualization represents the spatial
arrangement of hespdiv polygons within the territory in 2D.
}
\note{
A much clearer way to visualize polygons is by using the \code{blok3D}
function, with \code{height = "rank"}. However, a 3D plot is less suitable
option for papers.
}
\examples{
poly_scheme(example_hespdiv)

}
\seealso{
Other HespDiv visualization options: 
\code{\link{blok3d}()},
\code{\link{create_gif}()},
\code{\link{dendro}()},
\code{\link{plot.nullhespdiv}()},
\code{\link{plot_cs_hsa}()},
\code{\link{plot_hespdiv}()},
\code{\link{plot_hsa}()},
\code{\link{plot_hsa_q}()},
\code{\link{polypop}()}
}
\author{
Liudas Daumantas
}
\concept{HespDiv visualization options}
