% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_hespdiv.R
\name{print.hespdiv}
\alias{print.hespdiv}
\title{Print a hespdiv object}
\usage{
\method{print}{hespdiv}(x, ...)
}
\arguments{
\item{x}{A \code{hespdiv} object.}

\item{...}{Additional arguments, currently ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Formats and prints the rounded \code{split.stats} data frame from a
\code{hespdiv} object.
}
\author{
Liudas Daumantas
}
