% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_nullhespdiv.R
\name{print.nullhespdiv}
\alias{print.nullhespdiv}
\title{Print the results of nullhespdiv object}
\usage{
\method{print}{nullhespdiv}(x, ...)
}
\arguments{
\item{x}{A nullhespdiv class object}

\item{...}{other arguments}
}
\value{
x
}
\description{
print method for nullhespdiv class object
}
\author{
Liudas Daumantas
}
