% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block3D.R
\name{blok3d}
\alias{blok3d}
\title{Draw hespdiv polygons in 3D space}
\usage{
blok3d(
  obj,
  height = "mean",
  color.seed = 1,
  lines = TRUE,
  pnts.col = NULL,
  obs = TRUE
)
}
\arguments{
\item{obj}{An object of the hespdiv class}

\item{height}{A character vector with default value \code{"mean"}. Determines
which information from the \code{poly.stats} data frame is encoded as the
height of the polygons. Options are \code{"mean"}, \code{"sd"},
\code{"best"}, \code{"z.score"}, \code{"str.best"},
\code{"str.z.score"}, and \code{"rank"}. Multiple values are allowed.}

\item{color.seed}{An integer that controls the colors of the polygons. Change
it to a different number if you want to get a different set of colors.}

\item{lines}{A Boolean value. Do you want split-lines to be displayed over
the top of the polygons?}

\item{pnts.col}{A character or numeric vector. Color codes to be used for
displaying observations.}

\item{obs}{A Boolean value. Do you want observations to be displayed over the top of
the polygons?}
}
\value{
No return value. Called for the side effect of producing a 3D plot
using the \pkg{rgl} graphics engine.
}
\description{
This function visualizes HespDiv polygons in 3D space. The height axis
corresponds to a chosen column from the "poly.stats" data frame.
}
\details{
The function opens an \pkg{rgl} device for each selected column of the
\code{poly.stats} data frame.

Visualizing values from \code{poly.stats} as polygon height can provide
insight into spatial heterogeneity in the analyzed data and its hierarchical
spatial structure.

The \code{height = "rank"} option provides a more intuitive way to understand
the position of each polygon in the spatial hierarchy than
\code{poly_scheme()}.

Because higher-rank polygons are displayed above lower-rank polygons, they may
obscure the view. For this reason, \code{polypop(obj, height)} can be used
with the same arguments to interactively select unwanted polygons and remove
them from the actives plot.
}
\seealso{
Other HespDiv visualization options: 
\code{\link{create_gif}()},
\code{\link{dendro}()},
\code{\link{plot.nullhespdiv}()},
\code{\link{plot_cs_hsa}()},
\code{\link{plot_hespdiv}()},
\code{\link{plot_hsa}()},
\code{\link{plot_hsa_q}()},
\code{\link{poly_scheme}()},
\code{\link{polypop}()}
}
\author{
Liudas Daumantas
}
\concept{HespDiv visualization options}
