% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block3D.R
\name{blok3d}
\alias{blok3d}
\title{Draw hespdiv polygons in 3D space}
\usage{
blok3d(
  obj,
  height = "mean",
  color.seed = 1,
  lines = TRUE,
  pnts.col = NULL,
  obs = TRUE
)
}
\arguments{
\item{obj}{hespdiv object}

\item{height}{character (default 'mean'). What information from poly.stats
data frame do you wish to encode as the height of the polygons? Options:
"mean", "sd", "best", "z.score", "str.best", "str.z.score". See details
for explanations.}

\item{color.seed}{integer. Controls the color of polygons. Change it to a
different number if you want to get a different set of colors.}

\item{lines}{logical. Do you want split-lines to be displayed over the top of
the polygons?}

\item{pnts.col}{vector. Color codes to be used for displaying observations.}

\item{obs}{logical. Do you want observations to be displayed over the top of
the polygons?}
}
\description{
This function vizualise polygons obtained by hespdiv function. Polygon height
(z coordinate) encodes chosen information from the poly.stats data frame.
}
\author{
Liudas Daumantas
}
