% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gif.R
\name{create_gif}
\alias{create_gif}
\title{Create a gif of blok3D}
\usage{
create_gif(
  output_file = "rotating_polygons.gif",
  frames = 90,
  angle_per_frame = 5,
  fps = 10
)
}
\arguments{
\item{output_file}{name and path of output file}

\item{frames}{integer. Number of frames.}

\item{angle_per_frame}{numeric. Angle to rotate per frame in degrees.}

\item{fps}{integer. Frames per second.}
}
\description{
Functions creates a gif of rotating polygons displayed in a currently actively rgl device produced with blok3D
}
\seealso{
Other HespDiv visualization options: 
\code{\link{dendro}()},
\code{\link{poly_scheme}()}
}
\author{
Liudas Daumantas
}
\concept{HespDiv visualization options}
