% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_comp.R
\name{cross_comp}
\alias{cross_comp}
\title{Calculate polygon cross-comparison matrix}
\usage{
cross_comp(obj)
}
\arguments{
\item{obj}{hespdiv object}
}
\description{
This function calculates cross-comparison matrix of polygons identified by
hespdiv. This matrix can be transformed into distance matrix to be used in
cluster analysis.
}
\author{
Liudas Daumantas
}
