% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro.R
\name{dendro}
\alias{dendro}
\title{Plot hespdiv dendrogram}
\usage{
dendro(
  obj,
  poly.scheme = NULL,
  color = 1,
  performance.col = "blue",
  labels.col = 1,
  offset.factor = 1,
  arrange = TRUE,
  grob = TRUE,
  label.size = 0.5
)
}
\arguments{
\item{obj}{A hespdiv object.}

\item{poly.scheme}{ggplot2 object produced with poly_scheme function. Provide if you want identical colors for polygons in both plots.}

\item{color}{color vector used for dendrogram nodes and branches.}

\item{performance.col}{color vector used for text, displaying difference values between polygons}

\item{labels.col}{color vector used for text, displaying polygon IDs.}

\item{offset.factor}{numeric value used to scale the offset distance of displayed polygon IDs and performance values from a dendrogram node. Adjust experimentally, if you don't like the current distance.}

\item{arrange}{logical. Plot hespdiv dendrogram above polygon scheme?}

\item{grob}{logical. Convert plot to grob? Must be true, if you want to arrange polygon scheme and the dendrogram in a single plot.}

\item{label.size}{size of labels.}
}
\value{
A \code{grob} or \code{TableGrob} object if \code{grob = TRUE}, otherwise \code{NULL}.
}
\description{
This function displays a dendrogram of polygons produced by hespdiv split-lines. Branch length is proportional to difference. If performance of split-lines is a similarity measure, it is internally converted to difference.
}
\note{
If you want to transform similarity to difference externally, before applying dendro, change maximize to TRUE in the call info of obj.
}
\examples{
scheme <- poly_scheme(example_hespdiv)
# spatial visualization of polygons:
scheme
# dendrogramic visualization of polygons, using colors from 'scheme':
dendro(example_hespdiv, poly.scheme = scheme, arrange = FALSE, grob = FALSE)
}
\seealso{
Other HespDiv visualization options: 
\code{\link{blok3d}()},
\code{\link{create_gif}()},
\code{\link{plot.nullhespdiv}()},
\code{\link{plot_cs_hsa}()},
\code{\link{plot_hespdiv}()},
\code{\link{plot_hsa}()},
\code{\link{plot_hsa_q}()},
\code{\link{poly_scheme}()},
\code{\link{polypop}()}
}
\author{
Liudas Daumantas
}
\concept{HespDiv visualization options}
