% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data points that lie inside a polygon}
\usage{
get_data(polygon, data)
}
\arguments{
\item{polygon}{A data frame of 2 columns (x,y), containing coordinates of polygon vertices}

\item{data}{A data frame, containing "X" and "Y" columns. Other columns may also be present.}
}
\value{
A filtered data frame
}
\description{
Function allows to extract points from a provided dataset that lie inside a given polygon
}
\note{
This function drops points that lie strictly outside of a given polygon. If a point lies on a relative interior of an edge
of polygon (if polygon is open) or point is a vertex of polygon, than it will be retained.
}
\examples{
#Creating data frame of a polygon
poly<- data.frame(c(3.38,3.30,1.70,0.78,-0.06,-2.30,-2.94,-3.97,-1.61,-0.39,0.68,1.28,1.60,3.38),
c(-0.12,-0.31,-2.73,-3.22,-3.29,-2.19,-1.62,0.94,3.10,3.00,2.91,2.49,2.20,-0.12))

#Creating a dataset of points
data<-data.frame(X=runif(250,-4,4),Y=runif(250,-4,4))
plot(poly,type='l',xlab="X",ylab="Y")
points(data)

#Extracting points that lie inside a polygon
points(get_data(poly,data),pch=19,col=2)
}
\references{
Uses the C function InPoly(). InPoly is Copyright (c) 1998 by Joseph O'Rourke. It may be freely redistributed in its entirety provided that this copyright notice is not removed.
}
\author{
Liudas Daumantas
}
