% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hespdiv.R
\name{ggplot_hespdiv}
\alias{ggplot_hespdiv}
\title{Plot the results of hespdiv object}
\usage{
ggplot_hespdiv(
  obj,
  data,
  type = "color",
  n.loc = FALSE,
  pnts.col = NULL,
  seed = 10,
  ...
)
}
\arguments{
\item{obj}{hespdiv object}

\item{data}{data points used in hespdiv.}

\item{type}{character string. Either "width" or "color" (default "color").
Determines whether quality of split-lines is expressed by line width or
color.}

\item{n.loc}{logical. Would you like to visualize the number of observations
at each location? Only possible, when there are localities with more than
one observation. If type is 'color', then no. of obs. are expressed via point
sizes. Otherwise, they are expressed by color in logarithmic scale.}

\item{pnts.col}{color of data points.}

\item{seed}{integer. Seed used to randomized the colors of the split-lines.
Only meaningful, when argument type is 'width'.
Try setting different value, if colors of parallel split-lines or nearby
labels look too similar or to increase the general appeal of the graph.}

\item{...}{other arguments}
}
\value{
ggplot object
}
\description{
Function prints the results of hespdiv object.
}
\author{
Liudas Daumantas
}
