% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_effect.R
\name{group_effect}
\alias{group_effect}
\title{Investigate group effect}
\usage{
group_effect(obj, group, ...)
}
\arguments{
\item{obj}{A \code{hespdiv} object containing all necessary components (data, XY coordinates, etc.).}

\item{group}{A factor (or coercible to factor) defining the group of each observation in \code{obj$call.info$Call_ARGS$data}.}

\item{...}{Additional arguments passed to \code{\link{plot_hespdiv}}.
(\code{obj}, \code{performance}, and \code{subtitle} are preset within this function.)}
}
\value{
A list containing two elements:
\itemize{
\item \code{comp.vals}: A matrix of size \code{(number_of_splits) x (number_of_groups)},
giving the recomputed performances for each split-line and each group.
\item \code{plots}: A list of the \code{plot_hespdiv} outputs (one for each group).
}
}
\description{
Re-evaluates the performance of split-lines within a \code{hespdiv} object,
separately for each subset defined by a grouping factor.

For each level of \code{group}, the function:
\enumerate{
\item Subsets the data and coordinates to observations belonging to that group
\item Recomputes the performance measure (using \code{compare.f} and \code{generalize.f})
for each split-line in \code{obj}
\item Updates \code{obj$split.stats$performance} for that group
\item Calls \code{plot_hespdiv()}, storing the resulting plot (with a group-specific subtitle)
}
}
\examples{
\dontrun{
  # Suppose we have a hespdiv object called hd_obj and a grouping factor g:
  result <- group_effect(hd_obj, g, n.loc = TRUE)
  # result$comp.vals is the matrix of performance measures
  # result$plots is a list of ggplot/plot objects for each group
}
}
\seealso{
Other functions for hespdiv post-prossesing: 
\code{\link{cross_comp}()},
\code{\link{hsa_detailed}()},
\code{\link{hsa_sample_constrained}()},
\code{\link{hsa}()}
}
\concept{functions for hespdiv post-prossesing}
