% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hespdiv_nullltest.R
\name{nulltest}
\alias{nulltest}
\title{Test significance of spatial structure in data}
\usage{
nulltest(obj, n)
}
\arguments{
\item{obj}{An object of class \code{hsa}}

\item{n}{Number of permutations}
}
\value{
A "nullhespdiv" class object. It is a list containing two elements:
\itemize{
\item{1)}{ A data.frame which reports the obtained mean and standard deviation of split-line
performances after data shuffling. Also, it compares these performances with
the ones estimated with hespdiv analysis by providing a quantiles and z-scores.}
\item{2)}{ A data frame with all obtained comparison values for each split-line after data shuffling.}
}
}
\description{
Function n times shuffles the data and each time
checks what is the performance of each split-line dividing the new data subsets.
}
