% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nullhespdiv.R
\name{plot.nullhespdiv}
\alias{plot.nullhespdiv}
\title{Plot the results of nullhespdiv object}
\usage{
\method{plot}{nullhespdiv}(x, ...)
}
\arguments{
\item{x}{A nullhespdiv class object}

\item{...}{other arguments}
}
\description{
plot method for nullhespdiv class object
}
\seealso{
Other HespDiv visualization options: 
\code{\link{create_gif}()},
\code{\link{dendro}()},
\code{\link{plot_cs_hsa}()},
\code{\link{plot_hespdiv}()},
\code{\link{plot_hsa_q}()},
\code{\link{plot_hsa}()},
\code{\link{poly_scheme}()}
}
\author{
Liudas Daumantas
}
\concept{HespDiv visualization options}
