% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nullhespdiv.R
\name{plot.nullhespdiv}
\alias{plot.nullhespdiv}
\title{Plot the results of nullhespdiv object}
\usage{
\method{plot}{nullhespdiv}(x, ...)
}
\arguments{
\item{x}{A nullhespdiv class object}

\item{...}{other arguments}
}
\description{
plot method for nullhespdiv class object
}
\author{
Liudas Daumantas
}
