% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hsa.R
\name{plot_hsa}
\alias{plot_hsa}
\title{Visualize the stability of hespdiv polygons}
\usage{
plot_hsa(
  obj,
  alpha = 0.6,
  split.col = "gray20",
  pnts.col = NULL,
  pol.col = "7",
  type = 1,
  basal.col = 2,
  max.lwd = 3,
  min.lwd = 0.5,
  split.col.seed = NULL,
  newplot = TRUE,
  seperated = TRUE
)
}
\arguments{
\item{obj}{An object of class \code{hsa}}

\item{alpha}{The alpha value for transparency of split lines. Default is 0.6.}

\item{split.col}{Color of alternative subdivision split-lines. Default is "gray20".}

\item{pnts.col}{The color of data points. Default is NULL.}

\item{pol.col}{The color of polygons. Default is "7".}

\item{type}{An integer indicating the type of plot. Default is 1.}

\item{basal.col}{The color of basal subdivision split-lines.}

\item{max.lwd}{The maximum line width for split-lines. Default is 3.}

\item{min.lwd}{The minimum line width for split-lines. Default is 0.5.}

\item{split.col.seed}{A seed for generating random colors for split lines. Default is NULL.}

\item{newplot}{Create a plot in new device?}

\item{seperated}{Boolean. When \code{type} is >= 3, open a new graphical device for each rank?}
}
\description{
The function uses hespdiv sensitivity analysis results to visually
demonstrate the stability of the basal hespdiv subdivision. This is achieved
by displaying both alternative and basal hespdiv subdivisions on the same plot.
}
\details{
The type parameter determines the type of plot to be generated:
\describe{
\item{type = 1:}{
Basic plot - Displays the alternative and basal hespdiv subdivisions on the
same plot without split-line ranks or titles.}
\item{type = 2:}{
 Plot with split-line ranks - Includes split-lines ranks in the plot.
 Each split-line is assigned a different line width based on its rank.}
\item{type = 3:}{
 Plot with separate ranks - Generates multiple plots, each representing
 split-line ranks up to a certain value.}
\item{type = 4:}{
 Plot with separate and isolated ranks - Similar to mode 3 but isolates the
 split-line ranks. Generates multiple plots, each representing a
 specific split-line rank.}
}

If the alternative subdivisions spatially converge, but the basal
subdivision lies far from the zone of convergence, then you can use
the \code{change_base} function to select a more representative alternative
subdivision to be used as the basal subdivision. However, you should check
if the arguments used in that subdivision makes sense.
}
\note{
'newplot' allows to correctly render legend in types 2 and 3, and lines
in general when drawing in an active device (use 'broom' otherwise to delete
devices).
}
\seealso{
Other {functions for hespdiv sensitivity analysis}: 
\code{\link{change_base}()},
\code{\link{hsa_detailed}()},
\code{\link{hsa_quant}()},
\code{\link{hsa}()},
\code{\link{plot_hsa_q}()}

Other {HespDiv visualization options}: 
\code{\link{blok3d}()},
\code{\link{plot_hespdiv}()},
\code{\link{plot_hsa_q}()},
\code{\link{poly_scheme}()}
}
\author{
Liudas Daumantas
}
\concept{{HespDiv visualization options}}
\concept{{functions for hespdiv sensitivity analysis}}
