% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_scheme.R
\name{poly_scheme}
\alias{poly_scheme}
\title{Schematic plot of hespdiv polygons}
\usage{
poly_scheme(obj, segment = TRUE, id = TRUE, seed = 1)
}
\arguments{
\item{obj}{hespdiv object}

\item{segment}{Boolean. Display the segments joining the centroids with
the split-lines?}

\item{id}{Boolean. Display the IDs of polygons?}

\item{seed}{Integer. Randomization seed to produce colors. Optimize by trial
and error, if too confusing colors appear.}
}
\description{
Produces a schematic visualization of subdivided territory, in which
the location of each polygon is made clear via their centroinds, their ID
label, and punctuated lines that join polygon centroids with the split-lines
that produced that polygons.
}
