% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_pop.R
\name{polypop}
\alias{polypop}
\title{Remove polygons from rgl device}
\usage{
polypop(obj, height)
}
\arguments{
\item{obj}{The hespdiv object used to create the currently active rgl
device with the \code{blok3d} function.}

\item{height}{A character value that indicates the height co-ordinate.}
}
\description{
This function allows you to interactively select and remove unwanted polygons
from a 3D plot created with the \code{blok3d} function.
}
\seealso{
Other HespDiv visualization options: 
\code{\link{blok3d}()},
\code{\link{create_gif}()},
\code{\link{dendro}()},
\code{\link{plot.nullhespdiv}()},
\code{\link{plot_cs_hsa}()},
\code{\link{plot_hespdiv}()},
\code{\link{plot_hsa}()},
\code{\link{plot_hsa_q}()},
\code{\link{poly_scheme}()}
}
\author{
Liudas Daumantas
}
\concept{HespDiv visualization options}
