% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_pop.R
\name{polypop}
\alias{polypop}
\title{Remove polygons from rgl device}
\usage{
polypop(obj, height)
}
\arguments{
\item{obj}{The hespdiv object used to create the currently active rgl
device with the \code{blok3d} function.}

\item{height}{A character value that indicates the height co-ordinate.}
}
\description{
This function allows you to interactively select and remove unwanted polygons
from a 3D plot created with the \code{blok3d} function.
}
\seealso{
Other {function for hespdiv visualization in 3D}: 
\code{\link{blok3d}()}
}
\author{
Liudas Daumantas
}
\concept{{function for hespdiv visualization in 3D}}
