% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_pop.R
\name{polypop}
\alias{polypop}
\title{Remove polygons from rgl device}
\usage{
polypop(obj, first = TRUE, height)
}
\arguments{
\item{obj}{hespdiv object which was used to create currently active rgl
device with poly3d_hespdiv function.}

\item{first}{logical. Is this is a first call?}

\item{height}{character vector. Indicates what is the z coordinate.}
}
\description{
Function that allows interactively select and remove polygons from the graph.
}
