% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_hespdiv.R
\name{print.hespdiv}
\alias{print.hespdiv}
\title{Print the results of hespdiv object}
\usage{
\method{print}{hespdiv}(x, ...)
}
\arguments{
\item{x}{A hespdiv class object.}

\item{...}{other arguments}
}
\value{
x
}
\description{
Function formats and prints the rounded split.stats data frame.
from "hespdiv" class R object.
}
\author{
Liudas Daumantas
}
