% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_splits.R
\name{remove_splits}
\alias{remove_splits}
\title{Remove split-lines}
\usage{
remove_splits(obj, split.id, depend.splits = TRUE)
}
\arguments{
\item{obj}{hespdiv object}

\item{split.id}{vector of split-line ids}

\item{depend.splits}{logical. Remove split-lines that depend on specified
split-lines? If FALSE, only end-nodes of spatial dendrogram are removed.}
}
\value{
hespdiv object
}
\description{
Function returns hespdiv object, without split-lines of
specified id.
}
\examples{
## DO NOT RUN:
## Preparing fossil assemblage data:
## loading hespdiv data package that contains completed hespdiv analysis
## of fossil mammal occurrence data from US
# library(HDData)

## inspect hespdiv object
# plot_hespdiv(hd)

## identify weak splits
# weak_splits <- which(hd$split.stats$performance >= 0.3)
## remove weak splits
# performance_filtered <- remove_splits(obj = hd, split.id = weak_splits)
# plot_hespdiv(performance_filtered)

## inspect split-line significance
# nl
# plot(HHDATA::nl)
## identify non-significant split-lines
# nsig_splits <- which(nl[[1]]$quantile >= 0.05)
## remove the non-significant split-line
# sig_filtered <- remove_splits(obj = hd, split.id = nsig_splits)
# plot_hespdiv(sig_filtered)

## In case non-significant or weak split-line has good dependent split-lines,
## you may choose to retain them using depend.splits = FALSE.
## Attempt to remove 2nd split-line, that has dependent split-line (ID 3)
# unchanged_hd <- remove_splits(obj = hd, split.id = 2, depend = FALSE)
# plot_hespdiv(unchanged_hd)

## Alternatively, if you wish to remove non-significant or weak split-lines,
## even if they have good dependent split-lines, use depend.splits = TRUE.
## Attempt to remove 2nd split-line, that has dependent split-line (ID 3)
# changed_hd <- remove_splits(obj = hd, split.id = 2, depend = TRUE)
# plot_hespdiv(changed_hd)
}
